# packages/api-contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Goodgrocer API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/account/deletion": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Delete Account",
        "operationId": "delete_account_api_v1_account_deletion_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GoogleLoginInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GoogleLoginInput": {
        "properties": {
          "id_token": {
            "type": "string",
            "maxLength": 10000,
            "minLength": 1,
            "title": "Id Token"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id_token"
        ],
        "title": "GoogleLoginInput"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}